refactored to DuplexChannelHandler
